for #7529 new VMasonry layout component
